requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
Flask==2.2.2
Flask-PyMongo==2.3.0
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.1
